add arg to make empty suite deletion more explit in the code

+++ nac_test/robot_writer.py
@@ -161,7 +161,9 @@ class RobotWriter:
         path_parts = [output_path.name] + list(relative_path.parts) + [robot_file.stem]
         return ".".join([TestSuite.name_from_source(p) for p in path_parts if p])
 
-    def _update_ordering_entries(self, output_path: Path, robot_file: Path) -> None:
+    def _update_ordering_entries(
+        self, output_path: Path, robot_file: Path, delete_empty_suite: bool = False
+    ) -> None:
         """
         parse the resulting files and check if a) has at least one test case
         and b) if it has the "Test Concurrency" metadata set indicating that it
@@ -173,13 +175,15 @@ class RobotWriter:
         collector = TestCollector(full_suite_name)
         suite.visit(collector)
 
-        if len(collector.test_names) == 0:
+        if delete_empty_suite and len(collector.test_names) == 0:
             logger.info(
                 "Removing empty rendered robot file without any test cases: %s",
                 robot_file,
             )
             os.remove(robot_file)
-        elif collector.test_concurrency:
+            return
+
+        if collector.test_concurrency:
             logger.info(
                 "%s has been marked to be suitable for test concurrency, will run the tests in parallel",
                 robot_file,
@@ -276,13 +280,17 @@ class RobotWriter:
                                     str(output_path), rel, foldername, new_filename
                                 )
                             self.render_template(t_path, Path(o_path), env, **extra)
-                            self._update_ordering_entries(output_path, o_path)
+                            self._update_ordering_entries(
+                                output_path, o_path, delete_empty_suite=True
+                            )
                 if next_template:
                     continue
 
                 o_path = Path(output_path, rel, filename)
                 self.render_template(t_path, o_path, env)
-                self._update_ordering_entries(output_path, o_path)
+                self._update_ordering_entries(
+                    output_path, o_path, delete_empty_suite=True
+                )
 
         if ordering_file and len(self.ordering_entries) > 0:
             # sort the entries to keep the order by suite in the same way as robot/pabot would
